VOTE-DAPP/backend/requirements.txt
Flask
Flask-MySQLdb
Flask-Mail
python-dotenv
requests
sendgrid